+++ src/main/kotlin/org/kryptokrona/api/models/PostEncryptedGroup.kt
@@ -38,7 +38,7 @@ import java.time.LocalDateTime
 
 interface PostEncryptedGroup : Entity<PostEncryptedGroup> {
     companion object : Entity.Factory<PostEncryptedGroup>()
-    val id: Int
+    val id: Long
     var txHash: String
     var txSb: String
     var txTimestamp: Long
@@ -46,7 +46,7 @@ interface PostEncryptedGroup : Entity<PostEncryptedGroup> {
 }
 
 object PostEncryptedGroups : Table<PostEncryptedGroup>("postsencryptedgroup") {
-    val id = int("id").primaryKey().bindTo { it.id }
+    val id = long("id").primaryKey().bindTo { it.id }
     val txHash = varchar("tx_hash").bindTo { it.txHash }
     val txSb = varchar("tx_sb").bindTo { it.txSb }
     val txTimestamp = long("tx_timestamp").bindTo { it.txTimestamp }
